addAuthor mutation added
Fixed some variable names that were upper case, namely the types.
Still figuring out the best organization of files so right now the mongo-db module has a connect method used by the server and other methods used for data access or modification in graphql-schema

Just a thought for the future:
Maybe make a module with the resolvers and these are used in graphql-schema and the db access is only on the resolvers module.